upload

+++ app/component/add-buuton/add_button.tsx
@@ -6,6 +6,11 @@ interface add{
     id: number
     value: string
 }
+interface nam{
+    value: string
+    name: string
+    id: number
+}
 export default function Add_botton(){
     const [open, setOpen]= useState(false)
     
@@ -17,11 +22,15 @@ export default function Add_botton(){
         </>
     )
 }
-export function Delete_button({id,value}:add){
+
+export function Delete_button({id,value,name}:nam){
     const [open,setOpen] = useState(false)
     const [desable,setDisable]= useState(false)
+    const [confirm,setConfirm] = useState('')
+    const [cOpen,setCOpen] = useState(false)
     const handleDelete= async ()=> {
-        setDisable(true)
+        if(name === confirm){
+            setDisable(true)
             const {error:status_error } =  await supabase.from(value)
              .delete()
   .eq('id', id)
@@ -33,10 +42,17 @@ export function Delete_button({id,value}:add){
             setOpen(!open)
             setTimeout(() => {setOpen(false)   
             }, 2000);
+            setCOpen(false)
+        }
+        }
+        else {
+            console.error('error')
+            setCOpen(false)
+
         }
         }
     return(
-        <div className="relative">
+        <div className="static">
             {/* set save  */}
                 
                     {open ? (<div className="fixed top-0 right-0 bg-red-100 p-7 text-red-600 flex border border-red-500 rounded-lg">
@@ -50,11 +66,21 @@ export function Delete_button({id,value}:add){
 
         </div> ) : ''}
             {/*delete button */}
-             <button onClick={handleDelete} disabled={desable} className={`${desable ? "text-gray-500 text-[16px] size-10 pt-2 pl-2  text-center flex":"rounded-full text-red-500 text-[16px] hover:bg-red-900 hover:text-white size-10 pt-2 pl-2  text-center flex"}`} >
+             <button onClick={()=>setCOpen(true)} disabled={desable} className={`${desable ? "text-gray-500 text-[16px] size-10 pt-2 pl-2  text-center flex":"rounded-full text-red-500 text-[16px] hover:bg-red-900 hover:text-white size-10 pt-2 pl-2  text-center flex"}`} >
                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="size-6">
   <path strokeLinecap="round" strokeLinejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
 </svg>
 </button>
+        {/* confirm section */}
+        { cOpen ? <div className="fixed left-[30%] top-25 w-100 bg-gray-300 rounded-xl justify-items-center p-10 text-[20px]">
+            <div className="mt-10 mb-2">
+                Confirmer la suppression en écrivant ce nom <span className="font-bold">"{name}"</span>
+            </div>
+            <Input int={confirm} text="Confirmer"
+                                 type="text" out={(e)=>setConfirm(e.target.value)} require={false}/>
+             <button onClick={()=>setCOpen(false)} className="bg-gray-400 ml-5 rounded-2xl text-white text-[20px] hover:bg-gray-700 w-18.5 h-10">Cancel</button>
+            <button onClick={handleDelete} className="bg-[#2DAE0D] ml-3 rounded-2xl text-white text-[20px] hover:bg-green-700 w-18.5 h-10">delete</button>
+        </div>:null}
         </div>
     )
 }
